+++ examples/file-server/test/file-server.ix.ts
@@ -3,7 +3,7 @@ import { expect } from 'chai';
 import { FileServerDriver } from './file-server-driver';
 
 describe('File Server Feature', function() {
-    this.timeout(10_000);
+    this.timeout(15_000);
     const { getLoadedFeature } = withFeature({
         featureName: 'file-server/example',
         configName: 'file-server/run'
