+++ packages/scripts/src/index.ts
@@ -3,3 +3,4 @@ export * from './application';
 export * from './build-constants';
 export * from './types';
 export * from './node-environments-manager';
+export * from './forked-process';
